[WIP]: Fixed bug on SelledProductResponseDTO. Implemented stock management. Implemented derived query. Code improvements.

+++ src/main/java/ws/peoplefirst/point_of_sell/model/SelledProduct.java
@@ -36,7 +36,7 @@ public class SelledProduct {
         this.receipt = receipt;
         this.barCode = barCode;
         this.quantity = quantity;
-        this.total = singlePrice * quantity;
+        this.total = singlePrice * (double) quantity;
     }
 
     public UUID getId() {
